Add Request classes for conference validation and update controllers

+++ routes/web.php
@@ -10,30 +10,29 @@ use Inertia\Inertia;
 
 Route::get('/', function () {
     return Inertia::render('Main');
-});
+})->name('home');
 
-Route::get('/client/conferences', [ClientController::class, 'index']);
-Route::get('/client/conferences/{id}', [ClientController::class, 'show']);
-Route::post('/client/conferences/{id}/register', [ClientController::class, 'store']);
+Route::get('/client/conferences', [ClientController::class, 'index'])->name('client.conferences.index');
+Route::get('/client/conferences/{id}', [ClientController::class, 'show'])->name('client.conferences.show');
+Route::post('/client/conferences/{id}/register', [ClientController::class, 'store'])->name('client.conferences.register');
 
-Route::get('/employee/conferences', [EmployeeController::class, 'index']);
-Route::get('/employee/conferences/{id}', [EmployeeController::class, 'show']);
-
+Route::get('/employee/conferences', [EmployeeController::class, 'index'])->name('employee.conferences.index');
+Route::get('/employee/conferences/{id}', [EmployeeController::class, 'show'])->name('employee.conferences.show');
 
 Route::get('/admin', function () {
     return Inertia::render('Admin/Index');
-});
+})->name('admin.index');
 
-Route::get('/admin/users', [UsersController::class, 'index']);
-Route::get('/admin/users/{id}/edit', [UsersController::class, 'edit']);
-Route::put('/admin/users/{id}', [UsersController::class, 'update']);
+Route::get('/admin/users', [UsersController::class, 'index'])->name('admin.users.index');
+Route::get('/admin/users/{id}/edit', [UsersController::class, 'edit'])->name('admin.users.edit');
+Route::put('/admin/users/{id}', [UsersController::class, 'update'])->name('admin.users.update');
 
-Route::get('/admin/conferences', [ConferenceController::class, 'index']);
-Route::get('/admin/conferences/{id}/edit', [ConferenceController::class, 'edit']);
-Route::put('/admin/conferences/{id}', [ConferenceController::class, 'update']);
-Route::delete('/admin/conferences/{id}', [ConferenceController::class, 'destroy']);
+Route::get('/admin/conferences', [ConferenceController::class, 'index'])->name('admin.conferences.index');
+Route::get('/admin/conferences/{id}/edit', [ConferenceController::class, 'edit'])->name('admin.conferences.edit');
+Route::put('/admin/conferences/{id}', [ConferenceController::class, 'update'])->name('admin.conferences.update');
+Route::delete('/admin/conferences/{id}', [ConferenceController::class, 'destroy'])->name('admin.conferences.destroy');
 
-Route::get('/admin/conferences/create', [ConferenceController::class, 'create']);
-Route::post('/admin/conferences', [ConferenceController::class, 'store']);
+Route::get('/admin/conferences/create', [ConferenceController::class, 'create'])->name('admin.conferences.create');
+Route::post('/admin/conferences', [ConferenceController::class, 'store'])->name('admin.conferences.store');
